added function to validate rating before saving

+++ types.go
@@ -10,14 +10,14 @@ import (
 type Review struct {
 	Id        uuid.UUID `json:"id"`
 	MovieName string    `json:"movie_name"`
-	Rating    int       `json:"rating"`
+	Rating    int8      `json:"rating"`
 	Comment   string    `json:"comment"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
 type CreateReview struct {
 	MovieName string `json:"movie_name"`
-	Rating    int    `json:"rating"`
+	Rating    int8   `json:"rating"`
 	Comment   string `json:"comment"`
 }
 
@@ -27,7 +27,7 @@ type MyError struct {
 
 type APIFunc func(w http.ResponseWriter, r *http.Request) error
 
-func NewReview(movieName string, rating int, comment string) *Review {
+func NewReview(movieName string, rating int8, comment string) *Review {
 	return &Review{
 		Id:        uuid.New(),
 		MovieName: movieName,
